Move loopback specific and blockdevice general helpers to appropriate testtools modules.
