+++ flocker/node/agents/testtools/__init__.py
@@ -5,10 +5,26 @@ Test helpers for ``flocker.node.agents``.
 """
 
 from _cinder import (
-    make_icindervolumemanager_tests, make_inovavolumemanager_tests
+    make_icindervolumemanager_tests,
+    make_inovavolumemanager_tests,
 )
-
-
+from _blockdevice import (
+    detach_destroy_volumes,
+    make_iblockdeviceapi_tests,
+    mountroot_for_test,
+    umount,
+    umount_all,
+)
+from _loopback import (
+    loopbackblockdeviceapi_for_test,
+)
 __all__ = [
-    'make_inovavolumemanager_tests', 'make_icindervolumemanager_tests'
+    'detach_destroy_volumes',
+    'loopbackblockdeviceapi_for_test',
+    'make_iblockdeviceapi_tests',
+    'make_icindervolumemanager_tests',
+    'make_inovavolumemanager_tests',
+    'mountroot_for_test',
+    'umount',
+    'umount_all',
 ]
