+++ flocker/node/agents/testtools/_cinder.py
@@ -7,8 +7,9 @@ Test helpers for ``flocker.node.agents.cinder``.
 from zope.interface.verify import verifyObject
 
 from flocker.testtools import TestCase
-from .cinder import (
-    ICinderVolumeManager, INovaVolumeManager,
+from ..cinder import (
+    ICinderVolumeManager,
+    INovaVolumeManager,
 )
