+++ flocker/node/agents/testtools/_loopback.py
@@ -1,0 +1,35 @@
+# Copyright ClusterHQ Inc.  See LICENSE file for details.
+
+"""
+Test helpers for ``flocker.node.agents.blockdevice.LoopbackBlockDeviceAPI``.
+"""
+from os import getuid
+
+from ....testtools import random_name
+
+from ..loopback import LoopbackBlockDeviceAPI
+
+from . import detach_destroy_volumes
+
+
+def loopbackblockdeviceapi_for_test(test_case, allocation_unit=None):
+    """
+    :returns: A ``LoopbackBlockDeviceAPI`` with a temporary root directory
+        created for the supplied ``test_case``.
+    """
+    user_id = getuid()
+    if user_id != 0:
+        test_case.skipTest(
+            "``LoopbackBlockDeviceAPI`` uses ``losetup``, "
+            "which requires root privileges. "
+            "Required UID: 0, Found UID: {!r}".format(user_id)
+        )
+
+    root_path = test_case.mktemp()
+    loopback_blockdevice_api = LoopbackBlockDeviceAPI.from_path(
+        root_path=root_path,
+        compute_instance_id=random_name(test_case),
+        allocation_unit=allocation_unit,
+    )
+    test_case.addCleanup(detach_destroy_volumes, loopback_blockdevice_api)
+    return loopback_blockdevice_api
